serverutil/dirserver: separate GCP parts from dirserver command
Following the same pattern we used for cmd/upspinserver and
serverutil/upspinserver, this change moves the core functionality of
cmd/dirserver into the new serverutil/dirserver package.
When we move the GCP-related packages to the gcp.upspin.io repository,
the cmd/dirserver command in the core repository will remain without
the cloud parts.
