+++ cmd/dirserver/main.go
@@ -7,23 +7,12 @@
 package main // import "upspin.io/cmd/dirserver"
 
 import (
-	"flag"
-	"net/http"
-
 	"upspin.io/cloud/gcpmetric"
-	"upspin.io/cloud/https"
 	cloudLog "upspin.io/cloud/log"
-	"upspin.io/config"
-	"upspin.io/dir/inprocess"
-	"upspin.io/dir/server"
-	"upspin.io/errors"
-	"upspin.io/exp/dir/filesystem"
 	"upspin.io/flags"
 	"upspin.io/log"
 	"upspin.io/metric"
-	"upspin.io/rpc/dirserver"
-	"upspin.io/serverutil/perm"
-	"upspin.io/upspin"
+	"upspin.io/serverutil/dirserver"
 
 	// TODO: Which of these are actually needed?
 
@@ -42,10 +31,8 @@ const (
 	maxQPS        = 1000 // unlimited metric reports per second
 )
 
-var storeServerUser = flag.String("storeserveruser", "", "`user name` of the StoreServer")
-
 func main() {
-	flags.Parse(flags.Server, "kind", "project", "serverconfig")
+	flags.Register("project")
 
 	if flags.Project != "" {
 		cloudLog.Connect(flags.Project, serverName)
@@ -57,40 +44,5 @@ func main() {
 		}
 	}
 
-	// Load configuration and keys for this server. It needs a real upspin username and keys.
-	cfg, err := config.FromFile(flags.Config)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	// Create a new store implementation.
-	var dir upspin.DirServer
-	err = nil
-	switch flags.ServerKind {
-	case "inprocess":
-		dir = inprocess.New(cfg)
-	case "filesystem":
-		dir, err = filesystem.New(cfg, flags.ServerConfig...)
-	case "server":
-		dir, err = server.New(cfg, flags.ServerConfig...)
-	default:
-		err = errors.Errorf("bad -kind %q", flags.ServerKind)
-	}
-	if err != nil {
-		log.Fatalf("Setting up DirServer: %v", err)
-	}
-
-	// Wrap with permission checks, if requested.
-	var ready chan struct{}
-	if *storeServerUser != "" {
-		ready = make(chan struct{})
-		dir = perm.WrapDir(cfg, ready, upspin.UserName(*storeServerUser), dir)
-	} else {
-		log.Printf("Warning: no Writers Group file protection -- all access permitted")
-	}
-
-	httpDir := dirserver.New(cfg, dir, upspin.NetAddr(flags.NetAddr))
-	http.Handle("/api/Dir/", httpDir)
-
-	https.ListenAndServeFromFlags(ready, serverName)
+	dirserver.Main()
 }
